fix: start api gateway after otel bootstrap
